requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.4.0
frozenlist==1.4.1
gTTS==2.5.2
idna==3.7
multidict==6.0.5
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
